+++ tests/unit/validation/test_schema.py
@@ -5,6 +5,7 @@ import urllib.parse
 
 from dal.validation import Schema
 from dal.validation import SCHEMA_FOLDER_PATH
+from dal.classes.filesystem import FileSystem
 
 
 @pytest.fixture
@@ -15,28 +16,22 @@ def validator():
 
 
 @pytest.fixture
-def valid_data():
-    return {
-        "Label": "test",
-        "LastUpdate": {"user": "N/A", "date": "N/A"},
-        "Translations": {"en": "test", "pt": "teste", "es": "prueba"},
-    }
+def valid_data(metadata_folder):
+    data = FileSystem.read_json(metadata_folder / "Translation" / "delete_me.json")
+    return data["Translation"]["delete_me"]
 
 
 @pytest.fixture
-def invalid_data():
-    return {
-        "Label": "test",
-        "LastUpdate": {"user": "N/A", "date": "N/A"},
-        "Translations": "invalid_data",
-    }
+def invalid_data(metadata_folder_invalid_data):
+    data = FileSystem.read_json(metadata_folder_invalid_data / "Translation" / "delete_me.json")
+    return data["Translation"]["delete_me"]
 
 
 class TestSchema:
     def test_validate(self, validator, valid_data):
         """Test that valid data passes validation."""
         res = validator.validate(valid_data)
-        assert res["status"] is True
+        assert res["status"] is True, res["message"]
 
     def test_validate_negative(self, validator, invalid_data):
         """Test that invalid data fails validation."""